Create class Word, initial game set up
This commit adds a Word class with class methods set_up and
make_blank_spaces. The initialization takes two parameters, the word
from the txt file and how many characters that word has.

Once the Word object is created, it becomes the argument for the
play_game method which displays a greeting message and the number of
blanks for the user to see.

Currently, the blanks are printed before the message even though the
order is the other way around in the method. Another issue is that the
user's input is always a String type even if a number is typed in. This
means the if..elsif block isn't telling the user "not a letter", and
"Choose a letter".