Separate class methods and find_index method
This commit breaks apart the Word class method, set_up to be two
methods, blank_spaces and word. This was done to make it easier to
understand what happens when a new Word object is made. These two
class methods will also be two arguments for the play_game method.

The find_index method finds where the user's guess is within the word
array and returns the index number. This however, does not solve the
issue of whether there are multiple instances of the same letter
elsewhere in the word.

There have also been some variable name changes to make more sense.