Play game method and class method, blank_spaces
This commit makes the Word class method, blank_spaces, a couple lines
shorter by assigning the times loop to a variable in a single line.

The play_game method currently is able to identify the first occurance
of a matching guess. Currently working to swap the corresponding blank
space with the letter.