Word class attr_readers and update play_game
This commit removes the Word class methods, word and spaces. Rather than
have these methods, attr_readers for word and spaces were made.

Add change_spaces method which is used inside play_game to take three
arguments: spaces, index and guess. This method replaces the matching
dash at the index position with the guess and returns a string with the
results.

The play_game method now has a tries variable to keep track of how many
tries the user has left to complete the game. However, there are lots
of issues with introducing this variable. Currently there are a few
ways to fall into infinite loops that need to be sorted.

Some comments were removed to declutter the code.