Play_game and change_spaces methods
Currently beginning to work on being able to change multiple spaces if
the words contains a letter in different spaces. Within the play_game
method, if the user guesses a correct letter, the word_array gets put
through an each_with_index method with the letter and index as the
block variables. The each_with_index block will call the change_spaces
method if a letter of the word_array == guess.

If think the issue is that since a letter can be used multiple times in
a word, the change_spaces method is called that many times without
combining the two arrays together, putting out one long string.

*Regarding the previous commit, the issue of the infinite loop still
needs to be addressed.