Update change_spaces method
Change_spaces method's first two lines, creating a spaces_array object
and filling it with blanks is now moved inside play_game.

The change_spaces method can now detect if the user's guess is in more
than one spot in the word and swap the blank with that character. The
parameters of this method have been changed from (spaces, index, guess)
to (spaces, word, guess) where the the first two parameters are arrays
and the third is a number (the previous version's parameters were
number, number, string).