Transfer change_spaces method into new method
This commit takes the change_spaces method block and transfers it into
a new method, check_guess, and gets commented out. This is to try to
reduce the amount of code as well as make it easier for looping by
calling one method altogether instead of calling a method nested in
another.

The idea for the loop is to use an until loop - once the tries variable
reaches 0 or the user completes the word. The issue that I'm coming
across right now is that the result of line 103, check_guess(...), is a
NilClass and I can't seem to convert it to an array to compare against
word_array, which would part of my condition for the loop.