Loop inside play_game, check_guess method update
This commit adds a loop inside play_game method which starts with
decrementing the 'tries' var (this will have to be changed later to only
subtract when a guess is incorrect). The next block within the loop is
where the user's guess input is checked to be a valid input. The else
section in this block is where the check_guess method is called. The
loop's condition is when tries == 1 or when word_array == current (which
is a variable that is assigned the value of check_guess(...)).

The check_guess method has a wrong_letters array to store the user's
incorrect guesses throughout their turns. However, the array isn't
displaying all the wrong letters, only the latest one.

Other minor changes include moving a more detailed game introduction
message further up in the code and messages that are put out when the
user guesses correct or incorrect.