Delete find_index method, 'tries' decrement issue
This commit removes the find_index method because it is not necessary.

The 'tries' variable is not being properly decremented when a user makes
an incorrect guess. I think this might be because the first line in the
loop assigns 'tries' to 10, therefore, it is only at 9 no matter how
many incorrect guesses the user makes. The 'tries' variable may need to
be moved to another spot within play_game or within check_guess or be a
whole new method altogether.

One other issue related to the condition of the loop is that the loop
won't break after the word is complete. This has to do with unmatching
data types of word_array and the varible, current, which shows as a
NilClass. I'm having trouble understanding why the check_guess method
returns NilClass.

  *Initially, I thought it was returning nil because the 'current' array
  was blank but even after a few letters are filled in (or all are
  filled in), I still get nil.