Incorrect guesses array, 'wrong_letters'
This commit adds an array called wrong_letters used to store the user's
incorrect guesses.

The main issue with the previous version, when this variable was inside
the check_guess method, was that it would only display the latest wrong
guess from the user. This was because the array kept being reassigned to
an empty array everytime the method was called within play_game.

The 'wrong_letters' variable was moved inside play_game just outside of
the loop and it works properly (but still in need of  a few minor
adjustments).

Other changes in this commit are display-related.