Add correct_letters array and lose message
This commit prompts the user when they've made a guess that they've
already tried, whether it's correct or not. An array, correct_letters,
is initialized within play_game before the loop to take in letters that
are in the word_array.

Near the beginning of the loop inside play_game, a while loop checks
whether the user has guessed a correct or incorrect guess and prompt
for another guess if necessary.

If the user runs out of guess (tries == 0), the word is revealed along
with a 'you lose' message.