Add class Game, class methods
This commit adds a class Game along with two class methods, save and
load. This class has an attr_accessor to all relevant information to a
game once it has started (:word, :spaces, :solve_me, :correct_letters,
:wrong_letters). Currently the class methods don't contain much except
for YAML.dump(x) and YAML.load(x).

The idea that I'm working toward is to have an object of class Game
(most likely a hash) that has a key:value for each piece of information
in the game. When the user wants to save at the beginning of a turn, the
save method will be invoked, adding values (or rewriting values) in the
hash. When the user wants to load a previously saved game at the start
of a new game, the key:value pairs will load into the Game hash object.