Game class, save method
This commit adds a method within the Game class. A class variable, was
also added (@@hash = {}) to store all the values that are associated to
the current objects in the game.

The new Game object is created when the user types in 'save' in place of
a letter guess. A new Game object is created with the arguments (word,
spaces, correct_letters, wrong_letters). Initially I wanted 'solve_me'
as another argument but there is a scope issue with that value and it
may not be necessary. The save method creates keys with the same names
as the arguments and the values are the values of the variables inside
the play_game method.

Currently the Game.save method correctly puts the information on screen
like so:

  {"word"=>"analyses", "spaces"=>8, "correct_letters"=>["a", "e", "l",
  "n", "s"], "wrong_letters"=>["o", "p", "t", "r", "h", "g", "i"]}