Comment out Game class object, create File object
This commit bypasses the Game class object created in the last version.

I tried using File.new and File.open since this assignment calls for
serializing objects and the last version wasn't doing that. A new object,
'game_hash' is created with keys named for each piece of the game to
save but with no values to them.

When the user wants to save the game progress the 'game_hash' keys are
all assigned the current values of the game (word, spaces, correct_letters,
wrong_letters). Then the user is prompted to enter a filename which is
passed to 'save_file = File.new("#{x}.yaml", 'w+x')' to created a yaml
file within the directory with that name; if another file with the same
name exists, an error is raised. Afterward, the 'game_hash' is written
to the newly created file. Finally, the user exits the program.

Although the save function is working, creating a new yaml file in the
same directory as the hangman.rb file with the user assigned name, the
load feature is not loading up the saved game.