Add parameter, 'game_hash' to play_game
This commit adds another parameter to the play_game method. This is to
try and find a way to begin the method again once the user chooses to
load a previously saved game. I am trying to see if it's possible to
load in the saved game's 'word', 'correct_letters', 'tries' etc. as a
way to reassign those values to make it like the game is picking up from
where it was left off.

Currently the loading feature only shows a new, different game stats, as
opposed to one that was previously played on.