Clarify previous commit note
Re: 'Currently the loading feature only shows a new, different game
stats, as opposed to one that was previously played on.'

The save feature functions correctly and the load feature puts in the
correct information as far as the saved game. However, the saved game
stats do not overwrite the current new game. Typing in 'load' will
display a hash but start a new game.