Change loop, delete Game class, play_game method
This commit changes the loop type to:

  'while tries > 0 || word_array != spaces_array do

This was due to trying to utilize ways to break the loop when the 'load'
elsif block was called. This didn't end happening but the loop still
functions as it should so it was kept.

The Game class was also removed because it served no purpose.

The play_game method now has a lot of arguments: word, word_spaces,
word_array, spaces, spaces_array, tries, correct_letters, wrong_letters,
game_hash. This was to try to find a way to call 'play_game' inside the
loop after the user wants to load a saved game. Currently the issues are
still the same and unable to load properly.