Add hangman_v2.rb
This commit adds a different version of the assignment which uses a
Game class.

This version was created because I wanted to try using a class along
with class variables and methods throughout the code. Currently I'm
trying to get it to work as it did in the previous version.

The main obstacle with the previous version was trying to get it to load
a saved game. I haven't come across the same issue yet but I'm hoping
utilizing a class will help with that.