Play_game method, load feature
This commit doesn't add much, just noting the errors that occur. In the
previous commit, the error was due to 'no implicit conversion of String
into Integer'. This might've been due to data types being in the wrong
places with the correct_letters and/or wrong_letters arrays.

Currently the issue is that the loading feature brings in a sample yaml
file (that is recognized as a Hash) but when I check the class of
game_obj[:word] I get NilClass.

Another issue I will need to address later is how to create a document
in the correct YAML format with the game data in real time.