Play_game, load feature
This commit enables the user to load in a game from a sample YAML file
but only the word is recognized. Upon typing in 'load' and the filename,
the game has no pre-filled in correct letters, incorrect letters display
is empty and tries resets back to 10.

This was only after realizing that I needed to use a Game object as the
argument for play_game.