Loads in saved YAML but incomplete
This commit loads in a saved game but with some missing parts.

  *Only the incorrect letters are displayed on screen.
  *Number of tries left is correctly displayed on screen.
  *Correct letters are not filled in. They are also unable to be retried
  again.

When the loaded game ends, by running out of turns, another game is
immediately started and runs normally.

It is worth noting that the saved YAML file displays all the correct
information when the user decides to save.