User can load saved game
This commit allows the user to load in a previously saved game. The
correct_letters, wrong_letters and tries are all correct upon loading
in.

The reason for the correctly guessed letters not showing on the screen
was because 'spaces_array' is created as a blank array at the start of
the play_game method. To fix this, the correct_letters gets turned into
an array and word_array's elements are put through the each_with_index
enumerable method. The letters in the correct_letters array get filled
in to the spaces_array array at the index indicated in the word_array.

Currently the program still runs a new game after the user completes a
loaded game (whether by winning or losing).